Fix & clean
I remove the OV_RELEASE define in the lua file it's useless anyway, we never want feature in release and not in debug.
I Fix a log but for some reason it started to not compile anymore.
I Move the CoreLog definition in CoreModule.cpp otherwise too many file were definning CoreLog.