Implement a function that will give the path to the root folder of the engine
This was a bit more complicated that it sound, because I dont want to compute the path every time i need it, so I created a singleton class that hold on to a struct with static data.
Once I compute the path for the first time I store it in this static structure.
I also add to create a PlatformFileSystem class because we might not compute the path the same way on windows and other operating system.