Create a module specific build dll define
the thing is when your building a DLL that depends from another DLL
the DLL imported has to have his api to dll import and the dll being build has to have an api set to DLL export.
So by default API's are set to import, but when building a module we add the module specific define (OV_BUILD_CORE_DLL, OV_BUILD_ENGINE_DLL, ...) to switch the curent module api to DLL export.