Premake create helper functions
I created 3 helper function:
- UseDefines
- UseProjectDefaultConfig
- UseModuleDefaultConfig
Those helpers will be call in most of the project setups to set the config quickly.
Once you call one of those helpers you just have to specified what specific to this project, like the files, the include dirs, ...
I also modify the modules target output, all the dll are stored in a "Modules" directory and when building required dlls are copy next to the exe file.